Refactor, typo fixes, added new endpoints, added relationships, added Item entity, added OrderStatus enum, expanded functionality

+++ src/main/java/alkfejl_webshop/entity/BaseEntity.java
@@ -2,11 +2,7 @@ package alkfejl_webshop.entity;
 
 import lombok.Data;
 
-import javax.persistence.Column;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
+import javax.persistence.*;
 
 @Data
 @MappedSuperclass
@@ -15,5 +11,5 @@ public class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(nullable = false, updatable = false)
-    private Integer id;
+    private Long id;
 }
